+++ cmd/pgsql.go
@@ -16,12 +16,14 @@ var (
 		Usage:    "Run an interactive console with your PostgreSQL addon",
 		Flags: []cli.Flag{appFlag,
 			cli.StringFlag{Name: "size, s", Value: "", Usage: "Size of the container"},
+			cli.StringFlag{Name: "env, e", Value: "", Usage: "Environment variable name to use for the connection to the database"},
 		},
 		Description: ` Run an interactive console with your PostgreSQL addon.
 
    Examples
-    scalingo --app myapp pgsql-console
-    scalingo --app myapp pgsql-console --size L
+    scalingo --app my-app pgsql-console
+    scalingo --app my-app pgsql-console --size L
+    scalingo --app my-app pgsql-console --env MY_PSQL_URL
 
    The --size flag makes it easy to specify the size of the container executing
    the PostgreSQL console. Each container size has different price and performance.
@@ -31,14 +33,17 @@ var (
     # See also 'mongo-console' and 'mysql-console'
 `,
 		Action: func(c *cli.Context) {
-			currentApp := detect.CurrentApp(c)
-			opts := db.PgSQLConsoleOpts{
-				App:  currentApp,
-				Size: c.String("s"),
-			}
 			if len(c.Args()) != 0 {
 				cli.ShowCommandHelp(c, "pgsql-console")
-			} else if err := db.PgSQLConsole(opts); err != nil {
+				return
+			}
+
+			err := db.PgSQLConsole(db.PgSQLConsoleOpts{
+				App:          detect.CurrentApp(c),
+				Size:         c.String("s"),
+				VariableName: c.String("e"),
+			})
+			if err != nil {
 				errorQuit(err)
 			}
 		},