+++ cmd/redis.go
@@ -15,12 +15,14 @@ var (
 		Usage:    "Run an interactive console with your Redis addon",
 		Flags: []cli.Flag{appFlag,
 			cli.StringFlag{Name: "size, s", Value: "", Usage: "Size of the container"},
+			cli.StringFlag{Name: "env, e", Value: "", Usage: "Environment variable name to use for the connection to the database"},
 		},
 		Description: ` Run an interactive console with your Redis addon.
 
    Examples
-    scalingo --app myapp redis-console
-    scalingo --app myapp redis-console --size L
+    scalingo --app my-app redis-console
+    scalingo --app my-app redis-console --size L
+    scalingo --app my-app redis-console --env MY_REDIS_URL
 
    The --size flag makes it easy to specify the size of the container executing
    the Redis console. Each container size has different price and performance.
@@ -30,14 +32,17 @@ var (
     # See also 'mongo-console' and 'mysql-console'
 `,
 		Action: func(c *cli.Context) {
-			currentApp := detect.CurrentApp(c)
-			opts := db.RedisConsoleOpts{
-				App:  currentApp,
-				Size: c.String("s"),
-			}
 			if len(c.Args()) != 0 {
 				cli.ShowCommandHelp(c, "redis-console")
-			} else if err := db.RedisConsole(opts); err != nil {
+				return
+			}
+
+			err := db.RedisConsole(db.RedisConsoleOpts{
+				App:          detect.CurrentApp(c),
+				Size:         c.String("s"),
+				VariableName: c.String("e"),
+			})
+			if err != nil {
 				errorQuit(err)
 			}
 		},